nlp/requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
et-xmlfile==1.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.24.2
openai==0.27.4
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
sentence_transformers==2.2.2
google-cloud-speech==2.19.0
PyAudio==0.2.13
playsound==1.2.2
ffmpeg==1.4
google-auth==2.17.2
google-cloud==0.34.0
asyncio==3.4.3
asgiref==3.6.0
google-cloud-texttospeech==2.14.1
pydub==0.25.1